skillradar-ai/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy
psycopg2-binary
pydantic==2.6.4
python-dotenv>=1.0.0
spacy
beautifulsoup4
playwright
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
xgboost==2.0.3
sentence-transformers==2.7.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-experimental>=0.0.1
duckduckgo-search>=4.0.0
ddgs>=9.11.2
motor==3.3.2
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
